Update README.md
Road accidents are the most dreaded thing that can happen to a driver, despite the fact that they occur frequently. The majority of road users are aware of the general guidelines and safety precautions to take while on the road, but it is the road users' own negligence that causes accidents and wrecks. Human mistake is the leading cause of accidents and crashes. Despite of having rules and regulations on the roads, without the driver’s proper coordination and participation it seems imposible to lessen the fatalities. This paper aims to instigate everyone about road safety awareness through identifying the different risk that causes road accidents in region 4-A CALABARZON.